requirements.txt
django
djangorestframework
requests
Authlib